Refactor to core module
