+++ lifedb/__main__.py
@@ -1,6 +1,6 @@
 """Entrypoint script when running package with -m."""
 
-from lifedb import main
+from lifedb.run import run
 
 if __name__ == "__main__":
-    main()
+    run()
