+++ lifedb/core/__init__.py
@@ -1,0 +1,1 @@
+"""Core business logic for manipulating LifeDB data."""
