--- lifedb/lifedb.py
@@ -1,1 +1,0 @@
-"""Main module."""
