+++ pyproject.toml
@@ -117,7 +117,7 @@ requires = ["poetry-core>=1.0.0"]
 build-backend = "poetry.core.masonry.api"
 
 [tool.poetry.scripts]
-lifedb = "lifedb:main"
+lifedb = "lifedb.run:run"
 
 [tool.dagster]
 module_name = "lifedb.dagster.definitions"
